+++ lib/service_protocol/lib_client.rb
@@ -3,12 +3,12 @@
 require 'service_protocol/base_client'
 
 module ServiceProtocol
-  # Direct class protocol that bypasses ProxyAction/ValueObject.context serialization
-  # RemoteAction.call('action', params) === Action.call(params)
+  # Direct class protocol that bypasses Proxy/Entity.context serialization
+  # Remote.call('service_name:operation', params) === Operation.call(params)
   class LibClient < BaseClient
     def call
-      require 'service_protocol/proxy_action'
-      ServiceProtocol.constantizer(action).call(params)
+      require 'service_protocol/proxy'
+      ServiceProtocol.constantizer(operation).call(params)
     end
   end
 end