+++ spec/lib/service_protocol/proxy_spec.rb
@@ -1,3 +1,3 @@
 # frozen_string_literal: true
 
-# see remote_action_spec.rb
+# see remote_spec.rb