+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [1.1.1] - 2019-09-17
+
+### Changed
+- Renamed `Remote` to `Remote`, `Proxy` to `Proxy`, `Entity` to `Entity`.
+- Renamed `#action` to `#operation` throughout.
+- Removed `ruby_extensions` dependency.
+
 ## [1.1.0] - 2019-09-14
 
 ### Changed